Started the calibration program and commented out debug statement.
The calibration program is going to have to be interactive because the
findChessboardCorners does not perform well on heavily skewed
chessboards. The user will be prompted with an image and will have to
choose the correct chessboard points from the image. The user will be
able to zoom in and out to get a better picture of the corners. The
center will also be highlighted along with (hopefully) a line fit from
the chessboard corners to the center to show how close they align.