Got zooming in/out and adding/removing points working.
I might end up switching the control for zooming and points because its
really easy to accidentally add a point that may not be wanted. Should
also add a way to mark points on the image. Closing the window with the
window controls does not work and probably will not work since I need
GTK for that and setting that up as a pain in the ass.