Modified calibration image pipeline.
The rosparams that the image calibration script spit out are now saved
in a .yaml file in the launch_files directory.  When the specific load
line is called in the launch file, the rosparams for unwarping an image
will be loaded by default.  Now, there is no need to run the calibration
script on each load of 'roscore'.