Bounding line at start of course for mapper.
A new bounding line was created at the start of the map where the
robot is located.  This allows us to create a "fake" boundary to
force the robot to go in the correct location.  A boundary with a
probability of 100% is created that is 20 feet long, or the maximum
distance between two lanes.